Add files via upload
Coffee Chatbot
But first, coffee.

Whether it’s to get us ready to jump-start our day or to get us through a late-night cram session, many of us need our regular caffeine fix! Ordering coffee is just one example of a process that can be automated with the help of a chatbot.

You’re given the task of creating a Python chatbot that can help cut the wait time of a normal coffee run by taking customers’ orders in advance.